DDCNL-8950: Removed TaxAccountSummaryRepository and service refactoring
